updated README, added functions for fetching quotes from API when user shakes phone or clicks button
